requirements.txt
Django==1.10.4
django-appconf==1.0.2
django-mathfilters==0.4.0
django-pure-pagination==0.3.0
Django-Select2==5.8.10
MySQL-python==1.2.5
mysqlclient==1.3.9
requests==2.12.3
